+++ com.unity.cinemachine/Editor/Editors/CinemachineSequencerCameraEditor.cs
@@ -1,7 +1,3 @@
-#define USE_IMGUI_INSTRUCTION_LIST // We use IMGUI while we wait for UUM-27687 and UUM-27688 to be fixed
-
-using System.Collections.Generic;
-using System.Linq;
 using UnityEditor;
 using UnityEditor.UIElements;
 using UnityEngine;
@@ -35,30 +31,11 @@ namespace Unity.Cinemachine.Editor
                 HelpBoxMessageType.Info));
 
             var container = ux.AddChild(new VisualElement());
-#if USE_IMGUI_INSTRUCTION_LIST
-            // GML todo: We use IMGUI for this while we wait for UUM-27687 and UUM-27688 to be fixed
-            UpdateCameraCandidates();
-            container.Add(new IMGUIContainer(() =>
-            {
-                serializedObject.Update();
-                if (m_InstructionList == null)
-                    SetupInstructionList();
-                EditorGUI.BeginChangeCheck();
-                m_InstructionList.DoLayoutList();
-                if (EditorGUI.EndChangeCheck())
-                    serializedObject.ApplyModifiedProperties();
-            }));
-            container.TrackAnyUserActivity(() =>
-            {
-                if (targets.Length == 1)
-                    UpdateCameraCandidates();
-            });
-#else
             var vcam = Target;
             var header = container.AddChild(new VisualElement { style = { flexDirection = FlexDirection.Row, marginBottom = -2 } });
-            FormatInstructionElement(
-                header.AddChild(new Label("Child")), 
-                header.AddChild(new Label("  Blend in")), 
+            FormatInstructionElement(true,
+                header.AddChild(new Label("Child Camera")), 
+                header.AddChild(new Label("Blend in")), 
                 header.AddChild(new Label("Hold")));
             header.AddToClassList("unity-collection-view--with-border");
 
@@ -78,10 +55,9 @@ namespace Unity.Cinemachine.Editor
             list.makeItem = () => new BindableElement { style = { flexDirection = FlexDirection.Row }};
             list.bindItem = (row, index) =>
             {
-                // Remove children - items seem to get recycled
+                // Remove children - items get recycled
                 for (int i = row.childCount - 1; i >= 0; --i)
                     row.RemoveAt(i);
-                //row.parent.style.paddingLeft = row.parent.style.paddingRight = 0;
 
                 var def = new CinemachineSequencerCamera.Instruction();
                 var element = instructions.GetArrayElementAtIndex(index);
@@ -90,7 +66,14 @@ namespace Unity.Cinemachine.Editor
                 var vcamSel = row.AddChild(new PopupField<Object>());
                 vcamSel.formatListItemCallback = (obj) => obj == null ? "(null)" : obj.name;
                 vcamSel.formatSelectedValueCallback = (obj) => obj == null ? "(null)" : obj.name;
-                vcamSel.TrackAnyUserActivity(() => vcamSel.choices = Target.ChildCameras.Cast<Object>().ToList());
+                vcamSel.TrackAnyUserActivity(() => 
+                {
+                    vcamSel.choices ??= new();
+                    vcamSel.choices.Clear();
+                    var children = Target.ChildCameras;
+                    for (int i = 0; i < children.Count; ++i)
+                        vcamSel.choices.Add(children[i]);
+                });
         
                 var blend = row.AddChild(
                     new PropertyField(element.FindPropertyRelative(() => def.Blend), ""));
@@ -98,19 +81,20 @@ namespace Unity.Cinemachine.Editor
                     new InspectorUtility.CompactPropertyField(element.FindPropertyRelative(() => def.Hold), " "));
                 hold.RemoveFromClassList(InspectorUtility.kAlignFieldClass);
                     
-                FormatInstructionElement(vcamSel, blend, hold);
+                FormatInstructionElement(false, vcamSel, blend, hold);
 
-                vcamSel.BindProperty(vcamSelProp); // must bind at the end
+                // Bind must be last
                 ((BindableElement)row).BindProperty(element);
+                vcamSel.BindProperty(vcamSelProp);
             };
 
             // Local function
-            static void FormatInstructionElement(VisualElement e1, VisualElement e2, VisualElement e3)
+            static void FormatInstructionElement(bool isHeader, VisualElement e1, VisualElement e2, VisualElement e3)
             {
                 var floatFieldWidth = EditorGUIUtility.singleLineHeight * 3f;
                 
-                e1.style.marginLeft = 3;
-                e1.style.flexBasis = floatFieldWidth; 
+                e1.style.marginLeft = isHeader ? 2 * InspectorUtility.SingleLineHeight - 3 : 0;
+                e1.style.flexBasis = floatFieldWidth + InspectorUtility.SingleLineHeight; 
                 e1.style.flexGrow = 1;
                 
                 e2.style.flexBasis = floatFieldWidth; 
@@ -121,7 +105,6 @@ namespace Unity.Cinemachine.Editor
                 e3.style.flexGrow = 0;
                 e3.style.unityTextAlign = TextAnchor.MiddleRight;
             }
-#endif
             container.AddSpace();
             this.AddChildCameras(container, null);
 
@@ -135,102 +118,5 @@ namespace Unity.Cinemachine.Editor
 
             return ux;
         }
-
-#if USE_IMGUI_INSTRUCTION_LIST
-        UnityEditorInternal.ReorderableList m_InstructionList;
-        string[] m_CameraCandidates;
-        Dictionary<CinemachineVirtualCameraBase, int> m_CameraIndexLookup;
-
-        void OnEnable() => m_InstructionList = null;
-
-        void UpdateCameraCandidates()
-        {
-            List<string> vcams = new List<string>();
-            m_CameraIndexLookup = new Dictionary<CinemachineVirtualCameraBase, int>();
-            vcams.Add("(none)");
-            var children = Target.ChildCameras;
-            foreach (var c in children)
-            {
-                m_CameraIndexLookup[c] = vcams.Count;
-                vcams.Add(c.Name);
-            }
-            m_CameraCandidates = vcams.ToArray();
-        }
-
-        int GetCameraIndex(Object obj)
-        {
-            if (obj == null || m_CameraIndexLookup == null)
-                return 0;
-            var vcam = obj as CinemachineVirtualCameraBase;
-            if (vcam == null)
-                return 0;
-            if (!m_CameraIndexLookup.ContainsKey(vcam))
-                return 0;
-            return m_CameraIndexLookup[vcam];
-        }
-
-        void SetupInstructionList()
-        {
-            m_InstructionList = new UnityEditorInternal.ReorderableList(serializedObject,
-                    serializedObject.FindProperty(() => Target.Instructions),
-                    true, true, true, true);
-
-            // Needed for accessing field names as strings
-            var def = new CinemachineSequencerCamera.Instruction();
-
-            const float vSpace = 2f;
-            const float hSpace = 3f;
-            var floatFieldWidth = EditorGUIUtility.singleLineHeight * 2.5f;
-            var hBigSpace = EditorGUIUtility.singleLineHeight * 2 / 3;
-            m_InstructionList.drawHeaderCallback = (Rect rect) =>
-            {
-                float sharedWidth = rect.width - EditorGUIUtility.singleLineHeight
-                    - floatFieldWidth - hSpace - hBigSpace;
-                rect.x += EditorGUIUtility.singleLineHeight; rect.width = sharedWidth / 2;
-                EditorGUI.LabelField(rect, "Child");
-
-                rect.x += rect.width + hSpace;
-                EditorGUI.LabelField(rect, "Blend in");
-
-                rect.x += rect.width + hBigSpace; rect.width = floatFieldWidth;
-                EditorGUI.LabelField(rect, "Hold");
-            };
-
-            m_InstructionList.drawElementCallback = (Rect rect, int index, bool _, bool _) =>
-            {
-                if (m_CameraCandidates == null)
-                    return;
-                SerializedProperty instProp = m_InstructionList.serializedProperty.GetArrayElementAtIndex(index);
-                float sharedWidth = rect.width - floatFieldWidth - hSpace - hBigSpace;
-                rect.y += vSpace; rect.height = EditorGUIUtility.singleLineHeight;
-
-                rect.width = sharedWidth / 2;
-                SerializedProperty vcamSelProp = instProp.FindPropertyRelative(() => def.Camera);
-                int currentVcam = GetCameraIndex(vcamSelProp.objectReferenceValue);
-                int vcamSelection = EditorGUI.Popup(rect, currentVcam, m_CameraCandidates);
-                if (currentVcam != vcamSelection)
-                    vcamSelProp.objectReferenceValue = (vcamSelection == 0)
-                        ? null : Target.ChildCameras[vcamSelection - 1];
-
-                rect.x += rect.width + hSpace; rect.width = sharedWidth / 2;
-                if (index > 0 || Target.Loop)
-                    EditorGUI.PropertyField(rect, instProp.FindPropertyRelative(() => def.Blend),
-                        GUIContent.none);
-
-                if (index < m_InstructionList.count - 1 || Target.Loop)
-                {
-                    float oldWidth = EditorGUIUtility.labelWidth;
-                    EditorGUIUtility.labelWidth = hBigSpace;
-
-                    rect.x += rect.width; rect.width = floatFieldWidth + hBigSpace;
-                    SerializedProperty holdProp = instProp.FindPropertyRelative(() => def.Hold);
-                    EditorGUI.PropertyField(rect, holdProp, new GUIContent(" ", holdProp.tooltip));
-                    holdProp.floatValue = Mathf.Max(holdProp.floatValue, 0);
-
-                    EditorGUIUtility.labelWidth = oldWidth;
-                }
-            };
-        }
-#endif
     }
 }