+++ com.unity.cinemachine/Editor/Editors/CinemachineStateDrivenCameraEditor.cs
@@ -1,3 +1,4 @@
+#define USE_IMGUI_INSTRUCTION_LIST // We use IMGUI while we wait for UUM-27687 and UUM-27688 to be fixed
 #if CINEMACHINE_UNITY_ANIMATION
 using System;
 using System.Collections.Generic;
@@ -58,10 +59,13 @@ namespace Unity.Cinemachine.Editor
                 UpdateTargetStates();
                 layerSel.choices = m_LayerNames;
                 layerSel.SetValueWithoutNotify(m_LayerNames[layerProp.intValue]);
+#if USE_IMGUI_INSTRUCTION_LIST
                 UpdateCameraCandidates();
+#endif
                 noTargetHelp.SetVisible(Target.AnimatedTarget == null);
             });
             
+#if USE_IMGUI_INSTRUCTION_LIST
             // GML todo: We use IMGUI for this while we wait for UUM-27687 and UUM-27688 to be fixed
             UpdateTargetStates();
             UpdateCameraCandidates();
@@ -76,7 +80,9 @@ namespace Unity.Cinemachine.Editor
                 if (EditorGUI.EndChangeCheck())
                     serializedObject.ApplyModifiedProperties();
             }));
-
+#else
+            // GML todo: UITK implementation
+#endif
             ux.AddSpace();
             this.AddChildCameras(ux, null);
             this.AddExtensionsDropdown(ux);
@@ -216,6 +222,7 @@ namespace Unity.Cinemachine.Editor
             return m_StateIndexLookup[stateHash];
         }
 
+#if USE_IMGUI_INSTRUCTION_LIST
         void UpdateCameraCandidates()
         {
             List<string> vcams = new();
@@ -355,6 +362,7 @@ namespace Unity.Cinemachine.Editor
                     menu.ShowAsContext();
                 };
         }
+#endif
     }
 }
 #endif